clang-tidy

+++ julia.c
@@ -6,7 +6,7 @@
 /*   By: kostya <kostya@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/18 23:36:33 by kostya            #+#    #+#             */
-/*   Updated: 2021/09/23 15:02:39 by kostya           ###   ########.fr       */
+/*   Updated: 2021/09/23 15:42:38 by kostya           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,8 +32,9 @@ int	julia_1(const FLOAT *restrict field_real,
 		{
 			real = field_real[x];
 			imag = field_imag[y];
-			put_pixel(data, x, y, __recursive_color_compute_julia_extension(
-					data->params, real, imag));
+			put_pixel(data, (int)x, (int)y,
+				__recursive_color_compute_julia_extension(data->params,
+					real, imag));
 			++y;
 		}
 		++x;
@@ -64,5 +65,5 @@ inline int	__recursive_color_compute_julia_extension(
 			return (palette1((FLOAT)req / params->recursion_depth));
 		++req;
 	}
-	return (abs * 255);
+	return ((int)(abs * 255));
 }
